requirements.txt
asgiref==3.2.10
Django==3.0.10
django-cors-headers==3.5.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
phonenumbers==8.12.10
Pillow==7.2.0
python-barcode==0.13.1
pytz==2020.1
sqlparse==0.3.1